Compute the loss based on the combined tiles
Instead of computing the accuracy on a per-tile basis, compute it based
on the full board after all of the tiles has been re-combined. This is a
fair bit slower since Tensorflow was not really made for this type of
computation, but should give a much better policy since we are actually
judging what we will use.