Implement pattern embeddings
Add a 32 wide pattern embedding instead of the baseline features, these
embeddings are initialized to be the same as the baseline (with the
remaining 29 components being random).

This should allow the engine to forget the baseline patterns if it does
not feel they are useful, but also to keep them if they are useful.